Backend/requirements.txt
fastapi[all]
python-decouple
pymongo
